migrate log command (#358)
* migrate log command

* Update src/main/java/de/chojo/repbot/commands/log/handler/LogFormatter.java
