+++ src/main/java/de/chojo/repbot/ReputationBot.java
@@ -14,9 +14,9 @@ import de.chojo.repbot.commands.debug.Debug;
 import de.chojo.repbot.commands.locale.Locale;
 import de.chojo.repbot.commands.invite.Invite;
 import de.chojo.repbot.commands.info.Info;
-import de.chojo.repbot.commands.Log;
 import de.chojo.repbot.commands.prune.Prune;
 import de.chojo.repbot.commands.messages.Messages;
+import de.chojo.repbot.commands.log.Log;
 import de.chojo.repbot.commands.Reactions;
 import de.chojo.repbot.commands.repadmin.RepAdmin;
 import de.chojo.repbot.commands.RepSettings;
