+++ src/main/resources/locale.properties
@@ -107,15 +107,15 @@ command.locale.set.language.description=
 command.locale.set.message.invalidlocale=
 command.locale.set.message.set=
 command.log.description=
-command.log.donatedLog=
-command.log.messageLog=
-command.log.receivedLog=
-command.log.sub.donated=
-command.log.sub.donated.arg.user=
-command.log.sub.message=
-command.log.sub.message.arg.messageId=
-command.log.sub.received=
-command.log.sub.received.arg.user=
+command.log.donated.message.log=
+command.log.message.message.log=
+command.log.received.message.log=
+command.log.donated.description=
+command.log.donated.user.description=
+command.log.message.description=
+command.log.message.messageid.description=
+command.log.received.description=
+command.log.received.user.description=
 command.messages.description=
 command.messages.states.description=
 command.messages.states.message.choice.reactionConfirmation.false=
