+++ src/main/resources/locale_de_DE.properties
@@ -107,15 +107,15 @@ command.locale.set.language.description=Lege die neue Sprache fest.
 command.locale.set.message.invalidlocale=Diese Sprache existiert nicht.
 command.locale.set.message.set=Die Sprache wurde auf %LOCALE% gesetzt
 command.log.description=Erhalte Informationen über die Reputation eines Nutzers oder einer Nachricht.
-command.log.donatedLog=Vergebene Reputation von %USER%
-command.log.messageLog=Erhaltene Reputation der Nachricht %ID%
-command.log.receivedLog=Erhaltene Reputation von %USER%
-command.log.sub.donated=Sehe die vergebene Reputation eines Nutzers.
-command.log.sub.donated.arg.user=Der Benutzer, der die gespendete Reputation erhält.
-command.log.sub.message=Sehe die erhaltene Reputation für eine Nachricht.
-command.log.sub.message.arg.messageId=Die ID der Nachricht.
-command.log.sub.received=Sehe die erhaltene Reputation eines Nutzers.
-command.log.sub.received.arg.user=Der Benutzer, der den empfangenen Reputation erhalten soll.
+command.log.donated.message.log=Vergebene Reputation von %USER%
+command.log.message.message.log=Erhaltene Reputation der Nachricht %ID%
+command.log.received.message.log=Erhaltene Reputation von %USER%
+command.log.donated.description=Sehe die vergebene Reputation eines Nutzers.
+command.log.donated.user.description=Der Benutzer, der die gespendete Reputation erhält.
+command.log.message.description=Sehe die erhaltene Reputation für eine Nachricht.
+command.log.message.messageid.description=Die ID der Nachricht.
+command.log.received.description=Sehe die erhaltene Reputation eines Nutzers.
+command.log.received.user.description=Der Benutzer, der den empfangenen Reputation erhalten soll.
 command.messages.description=Einstellungen für die Benachrichtigung verwalten
 command.messages.states.description=Status der Benachrichtigungen verwalten
 command.messages.states.message.choice.reactionConfirmation.false=Der Bot sendet keine Benachrichtigung.
