+++ src/main/resources/locale_en_US.properties
@@ -107,15 +107,15 @@ command.locale.set.language.description=Define the new language.
 command.locale.set.message.invalidlocale=This language does not exist.
 command.locale.set.message.set=Language set to %LOCALE%.
 command.log.description=Get information about reputation of a user or message.
-command.log.donatedLog=Reputation donations of user %USER%.
-command.log.messageLog=Received Reputations for message %ID%.
-command.log.receivedLog=Received Reputations of user %USER%.
-command.log.sub.donated=Get the donated reputation of a user.
-command.log.sub.donated.arg.user=The user to get the dontated reputation.
-command.log.sub.message=Get the log for a message.
-command.log.sub.message.arg.messageId=The id of the message.
-command.log.sub.received=Get the received reputation of a user.
-command.log.sub.received.arg.user=The user to get the received reputation.
+command.log.donated.message.log=Reputation donations of user %USER%.
+command.log.message.message.log=Received Reputations for message %ID%.
+command.log.received.message.log=Received Reputations of user %USER%.
+command.log.donated.description=Get the donated reputation of a user.
+command.log.donated.user.description=The user to get the dontated reputation.
+command.log.message.description=Get the log for a message.
+command.log.message.messageid.description=The id of the message.
+command.log.received.description=Get the received reputation of a user.
+command.log.received.user.description=The user to get the received reputation.
 command.messages.description=Manage message settings
 command.messages.states.description=Manage states of messages
 command.messages.states.message.choice.reactionConfirmation.false=Bot will not send a message.
