+++ src/main/resources/locale_es_ES.properties
@@ -107,15 +107,15 @@ command.locale.set.language.description=Define el nuevo idioma.
 command.locale.set.message.invalidlocale=Este idioma no existe.
 command.locale.set.message.set=Establecer el idioma en %LOCALE%.
 command.log.description=Obtener información sobre la reputación de un usuario o mensaje.
-command.log.donatedLog=Donaciones de reputación del usuario %USER%.
-command.log.messageLog=Reputación recibida para el mensaje %ID%.
-command.log.receivedLog=Reptaciones recibidas del usuario %USER%.
-command.log.sub.donated=Obtener la reputación donada de un usuario.
-command.log.sub.donated.arg.user=El usuario para obtener la reputación donada.
-command.log.sub.message=Obtener el registro de un mensaje.
-command.log.sub.message.arg.messageId=El id del mensaje.
-command.log.sub.received=Obtener la reputación recibida de un usuario.
-command.log.sub.received.arg.user=El usuario para obtener la reputación recibida.
+command.log.donated.message.log=Donaciones de reputación del usuario %USER%.
+command.log.message.message.log=Reputación recibida para el mensaje %ID%.
+command.log.received.message.log=Reptaciones recibidas del usuario %USER%.
+command.log.donated.description=Obtener la reputación donada de un usuario.
+command.log.donated.user.description=El usuario para obtener la reputación donada.
+command.log.message.description=Obtener el registro de un mensaje.
+command.log.message.messageid.description=El id del mensaje.
+command.log.received.description=Obtener la reputación recibida de un usuario.
+command.log.received.user.description=El usuario para obtener la reputación recibida.
 command.messages.description=Gestionar la configuración de los mensajes
 command.messages.states.description=Gestionar los estados de los mensajes
 command.messages.states.message.choice.reactionConfirmation.false=El bot no enviará un mensaje.
