+++ src/main/resources/locale_fr_FR.properties
@@ -107,15 +107,15 @@ command.locale.set.language.description=Définit la nouvelle langue.
 command.locale.set.message.invalidlocale=Cette langue n'existe pas.
 command.locale.set.message.set=La langue est réglée sur %LOCALE%.
 command.log.description=Obtenir des informations sur la réputation d'un utilisateur ou d'un message.
-command.log.donatedLog=Réputation des dons de l'utilisateur %USER%.
-command.log.messageLog=Réputations reçues pour le message %ID%.
-command.log.receivedLog=Réception des réputations de l'utilisateur %USER%.
-command.log.sub.donated=Recevoir la réputation donnée d'un utilisateur.
-command.log.sub.donated.arg.user=L'utilisateur qui reçoit la réputation donnée.
-command.log.sub.message=Recevoir le journal d'un message.
-command.log.sub.message.arg.messageId=L'identifiant du message.
-command.log.sub.received=Obtient la réputation reçue d'un utilisateur.
-command.log.sub.received.arg.user=L'utilisateur pour obtenir la réputation reçue.
+command.log.donated.message.log=Réputation des dons de l'utilisateur %USER%.
+command.log.message.message.log=Réputations reçues pour le message %ID%.
+command.log.received.message.log=Réception des réputations de l'utilisateur %USER%.
+command.log.donated.description=Recevoir la réputation donnée d'un utilisateur.
+command.log.donated.user.description=L'utilisateur qui reçoit la réputation donnée.
+command.log.message.description=Recevoir le journal d'un message.
+command.log.message.messageid.description=L'identifiant du message.
+command.log.received.description=Obtient la réputation reçue d'un utilisateur.
+command.log.received.user.description=L'utilisateur pour obtenir la réputation reçue.
 command.messages.description=Gérer les paramètres des messages
 command.messages.states.description=Gérer les états des messages
 command.messages.states.message.choice.reactionConfirmation.false=Le robot n'enverra pas de message.
