+++ src/main/resources/locale_pt_PT.properties
@@ -107,15 +107,15 @@ command.locale.set.language.description=Definir o novo idioma.
 command.locale.set.message.invalidlocale=Esta língua não existe.
 command.locale.set.message.set=Língua definida a %LOCALE%.
 command.log.description=Recebe informação sobre a reputação de um utilizador ou mensagem.
-command.log.donatedLog=Reputação de doações de utilizador %USER%.
-command.log.messageLog=Reputações recebidas para mensagem %ID%.
-command.log.receivedLog=Reputações recebidas de utilizador %USER%.
-command.log.sub.donated=Abter a reputação doada de um utilizador.
-command.log.sub.donated.arg.user=O utilizador para obter a reputação doada.
-command.log.sub.message=Tem a reputação doada de um utilizador.
-command.log.sub.message.arg.messageId=O id da mensagem.
-command.log.sub.received=Abter a reputação recebida de um utilizador.
-command.log.sub.received.arg.user=O utilizador para obter a reputação recebida.
+command.log.donated.message.log=Reputação de doações de utilizador %USER%.
+command.log.message.message.log=Reputações recebidas para mensagem %ID%.
+command.log.received.message.log=Reputações recebidas de utilizador %USER%.
+command.log.donated.description=Abter a reputação doada de um utilizador.
+command.log.donated.user.description=O utilizador para obter a reputação doada.
+command.log.message.description=Tem a reputação doada de um utilizador.
+command.log.message.messageid.description=O id da mensagem.
+command.log.received.description=Abter a reputação recebida de um utilizador.
+command.log.received.user.description=O utilizador para obter a reputação recebida.
 command.messages.description=Gerir configurações de mensagens
 command.messages.states.description=Gerir estados de mensagens
 command.messages.states.message.choice.reactionConfirmation.false=O Bot não enviará uma mensagem.
