+++ src/main/resources/locale_ru_RU.properties
@@ -107,15 +107,15 @@ command.locale.set.language.description=Определи новый язык.
 command.locale.set.message.invalidlocale=Такого языка не существует.
 command.locale.set.message.set=Установлен язык: %LOCALE%.
 command.log.description=Получить информацию о репутации связанной с пользователем или сообщением.
-command.log.donatedLog=Выданные очки репутации пользователем %USER%.
-command.log.messageLog=Полученные очки репутации за сообщениe %ID%.
-command.log.receivedLog=Полученные очки репутации пользователя %USER%.
-command.log.sub.donated=Получить очки репутации выданные пользователем.
-command.log.sub.donated.arg.user=Пользователь для получения пожертвованной репутации.
-command.log.sub.message=Получить лог сообщений.
-command.log.sub.message.arg.messageId=Идентификатор сообщения.
-command.log.sub.received=Узнать очки репутации полученные пользователем.
-command.log.sub.received.arg.user=Пользователь для получения полученной репутации.
+command.log.donated.message.log=Выданные очки репутации пользователем %USER%.
+command.log.message.message.log=Полученные очки репутации за сообщениe %ID%.
+command.log.received.message.log=Полученные очки репутации пользователя %USER%.
+command.log.donated.description=Получить очки репутации выданные пользователем.
+command.log.donated.user.description=Пользователь для получения пожертвованной репутации.
+command.log.message.description=Получить лог сообщений.
+command.log.message.messageid.description=Идентификатор сообщения.
+command.log.received.description=Узнать очки репутации полученные пользователем.
+command.log.received.user.description=Пользователь для получения полученной репутации.
 command.messages.description=Управление настройками сообщений
 command.messages.states.description=Управление состояниями сообщений
 command.messages.states.message.choice.reactionConfirmation.false=Бот не будет отправлять сообщение.
